Walk, Run, Idle Animations and Flip Direction
Created walk, run, and idle animations and set them up in the Animator state machine to have transitions that go between each other when boolean animator properties are set to true or false as conditions. Also, wrote the code to flip a game object (in this case, the player) using local scale to also effect any other components like colliders on that game object to flip as well.